+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 * API:
     * Add `GET /admin/v2/users-csv/` (\#3276).
     * Fix handling of edge case upon updating user's `project_dirs` (\#3270).
+    * Add `slim` query parameter to `GET /api/v2/task-group/` (\#3277).
+    * Remove `args_schema` query parameter from `GET /api/v2/task-group/` (\#3277).
+    * Add `slim` query parameter to `GET /api/v2/task/` (\#3277).
+    * Remove `args_schema` query parameter from `GET /api/v2/task/` (\#3277).
     * Prevent setting non unique project names via project transfer (\#3281).
 * Internal:
     * Improve type hints for endpoint functions (\#3280).